// src/views/Stores/CreateMixture/interfaces.ts
export interface ModalTypes {
    URUNADI?: string;
    URUNRECETEURUNID?: string;
  }

  export interface IModalForm {
    URUNRECETEURUNID: string;
    URUNBIRIMID: string;
    DEPOID?: string;
    BIRIMID: string
    URUNID?: string
    TARIH?: string
    // DEPOID: string;
    // MIKTAR: number;
  }